Add List, Dict, and Map validators
List and Map validators both take a Validator object or list, and run
all values in the data value against those validators.

Dict validator takes in a mapping between key names and validator
objects or lists and runs only those specific key-value pairs against
those validators.

List and Map is good for a list of objects following a common type. Dict
is good for an assortment of various attributes on an object, which may
be different types, but must follow a strict format.